requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
camelot-py==0.9.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
clarifai==9.8.1
clarifai-grpc==9.8.2
click==7.1.2
coloredlogs==15.0.1
cryptography==41.0.3
dataclasses-json==0.5.14
diskcache==5.6.3
dotenv-python==0.0.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.3
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.35
googleapis-common-protos==1.60.0
grpcio==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kaleido==0.2.1
langchain==0.0.310
langsmith==0.0.43
llama-index==0.8.40
llama_cpp_python==0.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pdfminer.six==20221105
Pillow==9.5.0
pipdeptree==2.13.0
plotly==5.17.0
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pypdf==3.15.5
PyPDF2==2.12.1
PyPika==0.48.9
pyrate-limiter==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.11
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
reportlab==4.0.4
requests==2.31.0
rich==13.4.2
rpds-py==0.10.2
sec-edgar-downloader==5.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.64.1
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
validators==0.21.0
watchdog==3.0.0
watchfiles==0.20.0
weaviate-client==3.23.2
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
